+++ src/tile.rs
@@ -4,9 +4,9 @@ use sdl2::pixels::Color;
 use sdl2::render::Renderer;
 use sdl2::gfx::primitives::DrawRenderer;
 
-pub const INNER_TILE_SIZE: i16 = 14;
-pub const TILE_BORDER: i16 = 1;
-pub const TILE_SIZE: i16 = INNER_TILE_SIZE + TILE_BORDER * 2;
+pub const INNER_TILE_SIZE: u8 = 14;
+pub const TILE_BORDER: u8 = 1;
+pub const TILE_SIZE: u8 = INNER_TILE_SIZE + TILE_BORDER * 2;
 
 pub const HIDE_ROWS: u8 = 4;
 
@@ -18,10 +18,10 @@ pub fn draw_tile(renderer: &Renderer, pos: Pos, col: Color) {
 
     if y >= HIDE_ROWS as i16 {
         let y = y - HIDE_ROWS as i16;
-        let _ = renderer.box_(x * TILE_SIZE + TILE_BORDER,
-                              y * TILE_SIZE + TILE_BORDER,
-                              (x + 1) * TILE_SIZE - TILE_BORDER,
-                              (y + 1) * TILE_SIZE - TILE_BORDER,
+        let _ = renderer.box_(x * TILE_SIZE as i16 + TILE_BORDER as i16,
+                              y * TILE_SIZE as i16 + TILE_BORDER as i16,
+                              (x + 1) * TILE_SIZE as i16 - TILE_BORDER as i16,
+                              (y + 1) * TILE_SIZE as i16 - TILE_BORDER as i16,
                               col);
     }
 }
